Refactor simulation and display code to display subproject
