+++ settings.gradle
@@ -1,1 +1,1 @@
-include "robot", "display", "control", "common"
+include "common", "control", "robot", "display"
